+++ lambdasystemservice/pom.xml
@@ -93,6 +93,12 @@
             <artifactId>lambda-netty-loadbalancer</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.mongodb</groupId>
+            <artifactId>mongo-java-driver</artifactId>
+            <version>3.5.0</version>
+        </dependency>
+
 
 
     </dependencies>
