+++ lambdasystemservice/src/main/java/db/modal/FunctionDataModal.java
@@ -1,0 +1,16 @@
+package db.modal;
+
+import java.util.UUID;
+
+/**
+ * Created by deshan on 9/24/17.
+ */
+public class FunctionDataModal {
+    String domainName;
+    UUID instanceUuid;
+    String user;
+    String ipAddress;
+    String fileLocation;
+    FunctionType type;
+
+}
