+++ lambdasystemservice/src/main/java/db/modal/FunctionType.java
@@ -1,0 +1,11 @@
+package db.modal;
+
+/**
+ * Created by deshan on 9/24/17.
+ */
+public enum FunctionType {
+    JAVA,
+    PYTHON,
+    NODE_JS,
+    XML;
+}
