+++ lambdasystemservice/src/main/java/db/modal/ImageDataModal.java
@@ -1,0 +1,12 @@
+package db.modal;
+
+import java.util.UUID;
+
+/**
+ * Created by deshan on 9/24/17.
+ */
+public class ImageDataModal {
+    FunctionType type;
+    UUID imageUuid;
+
+}
